resolvido erro na comparacao de floats

+++ ExecutionEngine/ExecutionEngine.cpp
@@ -3504,7 +3504,7 @@ void ExecutionEngine::i_fcmpl(){
 	Valor valor2 = toppilha->desempilhaOperando();
 	Valor valor1 = toppilha->desempilhaOperando();
 	
-	int64_t num1, num2;
+	float num1, num2;
 	int32_t num3;
 	memcpy(&num1,&valor1.dado,4);
 	memcpy(&num2,&valor2.dado,4);
@@ -3535,7 +3535,7 @@ void ExecutionEngine::i_fcmpg(){
 	Valor valor2 = toppilha->desempilhaOperando();
 	Valor valor1 = toppilha->desempilhaOperando();
 	
-	int64_t num1, num2;
+	float num1, num2;
 	int32_t num3;
 	memcpy(&num1,&valor1.dado,4);
 	memcpy(&num2,&valor2.dado,4);
